requirements.txt
boto3
supabase
python-dotenv
requests
anthropic
opencv-python
pytesseract
Pillow
pdf2image
python-dotenv
mysql-connector-python
flask
numpy
PyMuPDF
pydantic-core
pydantic